rag/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.20

# Qdrant Vector Database
qdrant-client>=1.7.0

# Embeddings & LLM
sentence-transformers>=2.5.0
anthropic>=0.40.0
langchain==0.1.0
langchain-community==0.0.6

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Text Processing
# tiktoken==0.5.2  # Commented out due to Rust dependency
markdownify==0.11.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.12.0
flake8==6.1.0
